+++ README.md
@@ -3,5 +3,14 @@
 Relax brush for nodes, helps to arrange nodes easier.
 Shortcut: Shift+R
 
-Watch the video for more:
+Watch the video for demo:
 https://youtu.be/QvNz3ON6e1I
+
+##How to use the brush
+Controls:
+- Shift + R: invoke relax brush
+- Left click: smooth nodes
+- Right click or Escape key: finish
+- "[" to decrease brush size, "]" to increase it
+
+
